+++ paddle/fluid/eager/api/manual/eager_manual/forwards/CMakeLists.txt
@@ -1,10 +1,3 @@
-cc_library(
-  conv2d_fwd_function
-  SRCS conv2d_fwd_function.cc
-  DEPS ${eager_deps} ${fluid_deps} ${GLOB_OP_LIB} ${GLOB_OPERATOR_DEPS})
-
-add_dependencies(conv2d_fwd_function eager_codegen)
-
 set(eager_manual_functions
-    conv2d_fwd_function
+    ${PADDLE_SOURCE_DIR}/paddle/fluid/eager/api/manual/eager_manual/forwards/conv2d_fwd_function.cc
     PARENT_SCOPE)
