+++ paddle/fluid/eager/api/manual/fluid_manual/nodes/CMakeLists.txt
@@ -1,18 +1,5 @@
-cc_library(
-  fused_gate_attention_node
-  SRCS fused_gate_attention_node.cc
-  DEPS ${eager_deps} ${fluid_deps})
-
-cc_library(
-  fused_feedforward_node
-  SRCS fused_feedforward_node.cc
-  DEPS ${eager_deps} ${fluid_deps})
-
-cc_library(
-  fused_attention_node
-  SRCS fused_attention_node.cc
-  DEPS ${eager_deps} ${fluid_deps})
-
 set(fluid_manual_nodes
-    fused_gate_attention_node fused_feedforward_node fused_attention_node
+    ${PADDLE_SOURCE_DIR}/paddle/fluid/eager/api/manual/fluid_manual/nodes/fused_gate_attention_node.cc
+    ${PADDLE_SOURCE_DIR}/paddle/fluid/eager/api/manual/fluid_manual/nodes/fused_feedforward_node.cc
+    ${PADDLE_SOURCE_DIR}/paddle/fluid/eager/api/manual/fluid_manual/nodes/fused_attention_node.cc
     PARENT_SCOPE)
